requirements.txt
argparse==1.4.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
urllib3==2.5.0
gspread==6.2.1
google-auth-oauthlib==1.2.2
pytest==8.2.2